sua thanh gap 1 vat the thi chi tinh 1 lan

+++ PyMenu/func.py
@@ -2,12 +2,14 @@ import pygame
 
 from char import *
 
+
 class Func():
     def __init__(self):
         self.player_hp_last = 100
+        self.isCollideCoin = False
+        self.isCollideEnemy = False
+        # nhung cai
 
-        # nhung cai 
-        
     def ini_char(self):
         self.ADD_ENEMY = pygame.USEREVENT + 1
         pygame.time.set_timer(self.ADD_ENEMY, 250)
@@ -16,7 +18,6 @@ class Func():
         self.ADD_COIN = pygame.USEREVENT + 3
         pygame.time.set_timer(self.ADD_COIN, 1000)
 
-
         self.player = Player()
         self.enemies = pygame.sprite.Group()
         self.clouds = pygame.sprite.Group()
@@ -24,8 +25,10 @@ class Func():
         self.all_sprites = pygame.sprite.Group()
         self.all_sprites.add(self.player)
 
-        pygame.draw.rect(self.display, (255,0,0), pygame.Rect(self.DISPLAY_W - 300 - 10, 5, self.player_hp * 300 / self.player_hp_max, 20))
-        pygame.draw.rect(self.display, (255,0,0), pygame.Rect(self.DISPLAY_W - 300 - 10, 5, self.player_hp * 300 / self.player_hp_max, 20), 2)
+        pygame.draw.rect(self.display, (255, 0, 0),
+                         pygame.Rect(self.DISPLAY_W - 300 - 10, 5, self.player_hp * 300 / self.player_hp_max, 20))
+        pygame.draw.rect(self.display, (255, 0, 0),
+                         pygame.Rect(self.DISPLAY_W - 300 - 10, 5, self.player_hp * 300 / self.player_hp_max, 20), 2)
 
     def update_events(self):
         for event in pygame.event.get():
@@ -45,7 +48,6 @@ class Func():
                 self.coins.add(new_coin)
                 self.all_sprites.add(new_coin)
 
-
     def game_update(self):
         self.player.update(pygame.key.get_pressed())
         self.enemies.update()
@@ -54,12 +56,18 @@ class Func():
         for entity in self.all_sprites:
             self.window.blit(entity.image, entity.rect)
         if pygame.sprite.spritecollideany(self.player, self.coins):
-            self.player_score += 1
+            if not self.isCollideCoin:
+                self.player_score += 1
+                self.isCollideCoin = True
+        else:
+            self.isCollideCoin = False
+
         if pygame.sprite.spritecollideany(self.player, self.enemies):
-            self.player_hp -= 1
-            if self.player_hp <= 0:
-                self.player.kill()
-                self.finish_game()
-        
-
-
+            if not self.isCollideEnemy:
+                self.player_hp -= 1
+                self.isCollideEnemy = True
+                if self.player_hp <= 0:
+                    self.player.kill()
+                    self.finish_game()
+        else:
+            self.isCollideEnemy = False
